Add initial pass at string handling, no decoding yet.
This almost certainly needs more tests with other maps as well, as this
is part of where protections start doing shenanigans.

Initially I was trying to avoid eagerly parsing any of this stuff
(seemed better to avoid given that protections try to exploit anything
that is "more eager" than BW when playing games), but we need it for
basic stuff like the map title/description anyway. It will also be nice
to have this auto-detect the encoding of strings.