Add the beginnings of string decoding + encoding detection.
The logic for encoding detection is basically taken from bw-chk, with a
few fixes for things that appear to be (minor) logic bugs there. This
library is still missing parsing support for a number of things that
reference strings and are used in encoding detection in bw-chk, so those
will be added soon, and then I'll write more tests targeting the
detection (probably using bw-chk's test assets as well).