Add a map that requires using specific languages to extract the CHK file, also fix up a couple things that its protection showed were broken.
It should be noted that in SC:R, this map will be broken for any
language besides English that was available when BW was released (e.g.
French). Newer languages use English to extract MPQ files, so things
like Korean still work. Thus, this is an unfortunate Protector "trick",
as this will not work for users with that language set at all.