Restructure things slightly so Chk owns its data, add method for simple Chk extraction/parsing.
Chk taking ownership of its data makes things easier to deal with (since
Mpq returns an owned value, we'd need to return the Chk data as well
otherwise and it would need to be kept around for the lifetime of the
Chk).

I also had to update the StringsChunkData to keep a clone of its data as
well, as there were otherwise circular references to the overall data
(we needed to move it to create the Chk, but StringsChunkData was still
holding the reference). There are other options there as well (Rc, etc.)
but this seemed like the smaller change.